make annonation script standalone to simplify things
