--- .github/actions/annonate-browserstack/action.yml
@@ -1,13 +1,0 @@
-name: Annonate Browserstack
-description: Adds annonation to browserstack sessions' public url
-inputs:
-  browserstackArtifiactsDir:
-    description: directory containing json files with the sessions' public urls
-    required: true
-runs:
-  using: composite
-  steps:
-    - run: |
-        node index.js ${{ inputs.browserstackArtifiactsDir }}
-      working-directory: ./.github/actions/annonate-browserstack
-      shell: bash
